# projects/01-ai-knowledge-base/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  models.AuditDetails:
    additionalProperties: true
    type: object
  models.AuditLog:
    properties:
      action:
        type: string
      created_at:
        type: string
      details:
        $ref: '#/definitions/models.AuditDetails'
      duration_ms:
        type: integer
      error_message:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      request_method:
        type: string
      request_path:
        type: string
      resource_id:
        type: string
      resource_type:
        type: string
      status:
        type: string
      tenant_id:
        type: string
      user_agent:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  models.AuditLogStats:
    properties:
      success_rate:
        type: number
      top_actions:
        items:
          additionalProperties: true
          type: object
        type: array
      top_users:
        items:
          additionalProperties: true
          type: object
        type: array
      total_operations:
        type: integer
      unique_users:
        type: integer
    type: object
...
  models.UserAuditStats:
    properties:
      error_count:
        type: integer
      failure_count:
        type: integer
      success_count:
        type: integer
      top_actions:
        items:
          additionalProperties: true
          type: object
        type: array
      total_operations:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@saleschampionhub.com
    name: 技术支持
  description: "企业级AI知识库管理平台，支持多租户、RLS数据隔离和向量搜索功能。\n\n## \U0001F510 认证说明\n所有接口都需要JWT
    Token认证（从子项目0用户中心获取）。\n使用用户中心的 `/api/v1/auth/login` 接口获取Token。\n\n## ✨ 核心特性\n-
    多租户架构：租户/组织/用户三级隔离\n- Row-Level Security (RLS)：数据库级别的数据隔离\n- 三级挂载系统：租户 > 组织 >
    用户权限继承\n- 向量搜索：基于pgvector的语义搜索（1024维向量）\n- Redis多层缓存：热点数据缓存优化\n- 完整用户中心集成：统一认证授权"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: AI知识库管理平台 API
  version: "1.0"
paths:
  /api/v1/audit-logs:
    get:
      consumes:
      - application/json
      description: 查询审计日志，支持多种过滤条件
      parameters:
      - description: 租户ID
        in: query
        name: tenant_id
        type: string
      - description: 用户ID
        in: query
        name: user_id
        type: string
      - description: 操作类型
        in: query
        name: action
        type: string
      - description: 资源类型
        in: query
        name: resource_type
        type: string
      - description: 资源ID
        in: query
        name: resource_id
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 查询审计日志
      tags:
      - 审计日志
  /api/v1/audit-logs/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取审计日志详情
      parameters:
      - description: 审计日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuditLog'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取审计日志详情
      tags:
      - 审计日志
  /api/v1/audit-logs/export:
    get:
      consumes:
      - application/json
      description: 导出审计日志为CSV或JSON格式
      parameters:
      - description: 导出格式（csv/json）
        in: query
        name: format
        required: true
        type: string
      - description: 租户ID
        in: query
        name: tenant_id
        type: string
      - description: 用户ID
        in: query
        name: user_id
        type: string
      - description: 操作类型
        in: query
        name: action
        type: string
      - description: 资源类型
        in: query
        name: resource_type
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 导出审计日志
      tags:
      - 审计日志
  /api/v1/audit-logs/stats:
    get:
      consumes:
      - application/json
      description: 获取指定租户的审计统计信息
      parameters:
      - description: 统计天数（默认30天）
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuditLogStats'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取租户审计统计
      tags:
      - 审计日志
  /api/v1/audit-logs/users/{user_id}/stats:
    get:
      consumes:
      - application/json
      description: 获取指定用户的审计统计信息
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      - description: 统计天数（默认30天）
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserAuditStats'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取用户审计统计
      tags:
      - 审计日志
...
securityDefinitions:
  BearerAuth:
    description: 输入格式："Bearer " + JWT令牌（注意Bearer后面有空格）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"